requirements.txt
certifi==2023.5.7
clickhouse-connect==0.5.25
lz4==4.3.2
mysql-connector-python==8.0.33
numpy==1.24.3
protobuf==3.20.3
pytz==2023.3
urllib3==2.0.2
zstandard==0.21.0
